% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compliments.R
\name{provide_compliments}
\alias{provide_compliments}
\title{Provides compliments}
\usage{
provide_compliments()
}
\value{
character string
}
\description{
Provides compliments to QEARI peoples
}
\examples{
provide_compliments()
}
